refactor(explore_code): redesign sub-agent from first principles; fix prompt bugs
Sub-agent (src/.../tools/explore_code_subagent*):
- Split the 2.4k-line file into focused modules: candidates (evidence
  model, extraction, ranking, quote excerpting), report (schema, policy,
  rendering), prompts, and a slim orchestration entry.
- Model now only points at observed candidate IDs + narrative. Quotes are
  excerpted from observed source by code, so they cannot be fabricated
  (removes the quote-copy + verification + revision-round machinery).
- Replace the multi-stage action/confidence rewrite cascade with a single
  pure deriveOutcome(): action/confidence are functions of surviving
  evidence, so every invariant holds by construction.
- Replace the second "nudge" stream with prepareStep forcing submit_report
  on the final allowed step; deterministic fallback unchanged.
- Structured searchSuggestions replace the free-form query=... DSL string.
- Domain-neutral system prompt (no benchmark nouns); guard test expanded.
- Report's external format (Flow/Missing/Action + JSON paths) is unchanged,
  so the parent protocol and explore_code cache are untouched.
- Telemetry uses the structured outcome directly instead of re-parsing the
  rendered report. Dead fields (pathKinds, estimatedTokens) removed.

System prompt (src/prompts/local_agent_prompt.ts):
- Fix: SERVER_LAYER referred to a hardcoded "step 4" that is wrong when the
  app-blueprint step is prepended; now refers to the Implement step by name.
- Fix: $-unsafe String.replace on user-editable AI_RULES; use replacer fns.
- Move the recommendedPrimaryAction protocol into the explore_code tool
  description (single source of truth, rendered as an action table); the
  workflow step just points at it.
- Dedup the Pro/Basic development-workflow builders.
- Soften the per-edit mandatory re-read; generic no-write wording in Ask
  mode; allow tool flows that require user approval in tool-calling rule 5.
