+++ src/pro/main/ipc/handlers/local_agent/tools/explore_code.ts
@@ -87,10 +87,20 @@ export const exploreCodeTool: ToolDefinition<
   name: "explore_code",
   description: `Ask a code reconnaissance sub-agent to explore code included in a configured TypeScript project.
 
-Use this when you need to understand how a TypeScript, TSX, JavaScript, or JSX feature, symbol, type, component, service, or flow is implemented across files. It returns a compact flow with quoted evidence, relevant file/line ranges, read targets, confidence, and recommendedPrimaryAction. Treat a high- or medium-confidence report as the codebase map; follow recommendedPrimaryAction instead of rediscovering the code. If a high- or medium-confidence report says answer_from_report, answer or plan from the report. If it says read_targets, explain/locate requests should answer from the report and cite the listed targets as jump points without reading them, unless the report is low confidence or the user explicitly asks for edits, debugging, or exact source verification; edit/debug requests may read only the listed tight ranges before changing code or verifying exact source. If it says targeted_gap_search, run only the rendered Search targets with their exact terms/scopes, then answer from the report plus those search results and name any remaining gap; do not invent broader searches, open arbitrary hits, or call explore_code again. If no executable Search targets are rendered, answer from the observed flow while naming the remaining gap. If confidence is low, inspect the listed read targets or search targets before relying on the report.
-Use the intent argument to describe what you will do with the result: use explain for "trace how", data-flow, request-flow, or "how is this computed/surfaced" questions; use locate for finding the best files/symbols; use edit/debug when exact ranges will be read before changing code or verifying behavior.
+Use this when you need to understand how a TypeScript, TSX, JavaScript, or JSX feature, symbol, type, component, service, or flow is implemented across files. It returns a compact report: a Flow of file/line ranges with quoted evidence, optional Read targets and Search targets, a Confidence, and an Action. Treat a high- or medium-confidence report as the codebase map and follow its Action rather than rediscovering the code.
 
-Do not call this repeatedly for the same investigation after a high- or medium-confidence report. Use the report, or at most the exact rendered Search targets or tight edit/debug read targets.
+Set the intent argument to what you will do with the result: explain for "trace how", data-flow, request-flow, or "how is this computed/surfaced" questions; locate for finding the best files/symbols; edit or debug when you will read exact ranges before changing code or verifying behavior.
+
+Follow the report's Action:
+
+| Action | Do next | Do NOT |
+|--------|---------|--------|
+| answer_from_report | Answer or plan directly from the report. | Re-read discovery files, grep, or call explore_code again. |
+| read_targets | explain/locate: answer from the report, citing the listed targets as jump points. edit/debug: read only the listed tight ranges before changing or verifying code. | Read targets just to confirm the map for explain/locate (unless confidence is low or the user asked for edits/debugging/exact verification). |
+| targeted_gap_search | Run only the rendered Search targets with their exact terms/scopes, then answer from the report plus those results and name any remaining gap. | Invent broader searches, open arbitrary hits, or call explore_code again. |
+| skip_explore_result | Proceed without the report; nothing relevant was found. | Treat it as a map. |
+
+If confidence is low, inspect the listed read/search targets before relying on the report. If an Action calls for Search targets but none are rendered, answer from the observed Flow and name the remaining gap. Do not call explore_code repeatedly for the same investigation after a high- or medium-confidence report.
 
 Only use this for files included in the app's TypeScript config. JavaScript and JSX require TypeScript config support such as allowJs. If the project does not have TypeScript installed and configured, use grep/list_files/read_file instead.`,
   inputSchema: exploreCodeSchema,
