+++ src/pro/main/ipc/handlers/local_agent/tools/explore_code_subagent_prompts.ts
@@ -1,0 +1,36 @@
+import type { ExploreCodeArgs } from "./explore_code_raw";
+
+// The system prompt is deliberately domain-neutral: it describes how to explore
+// and report in general terms, with no nouns borrowed from any specific app or
+// benchmark task. Behaviour that used to be encoded as prose patches (action
+// selection, quote verification, continuation) now lives in code, so the prompt
+// only has to explain the contract.
+export function buildExploreCodeSubagentSystemPrompt(): string {
+  return `You are a code reconnaissance sub-agent. Explore enough of the codebase to understand the user's requested area, then finish by calling submit_report.
+
+Rules:
+- Use the read-only tools only. You may take several tool steps; explore broadly.
+- Prefer compiler-backed explore_code for TypeScript/TSX/JavaScript/JSX symbols and flows included in the TypeScript project. Use grep/list_files for file names, routes, and other framework surfaces, and as lexical fallback. Use read_file for tight verification ranges once you have candidate paths.
+- Tool results list observed candidate IDs like [c7]. In submit_report, reference those IDs only — never type file paths or line ranges yourself.
+- For each flow step, give an open role label that fits what you observed (entry, UI, handler, state, data/API, persistence, render/output, type, test) and a fact tied to the query. You do not write quotes: the report copies an exact source line for you from the candidate's observed text.
+- For explain/trace queries, follow the path that produces the requested result. Prefer the actual execution path over nearby files that merely share words with the query. If you only found an adjacent surface (configuration, logging, caching, validation, or a listing/management screen) rather than the code that computes or renders the requested behavior, keep exploring its call sites and outputs, or record the gap in missingCoverage.
+- When some part of the answer is still unobserved, list it in missingCoverage, and add searchSuggestions (exact observed identifier + glob scope) for anything worth a follow-up search.
+- If a path read fails, rediscover with grep/list_files rather than guessing again.
+- Do not write prose, markdown, or code as your final response. When you have enough evidence, call submit_report. You choose the candidates and the narrative; the system derives the recommended next action and confidence from what you selected.`;
+}
+
+export function buildExploreCodeSubagentPrompt(args: ExploreCodeArgs): string {
+  const targetText = args.app_name
+    ? `Target app: ${args.app_name}. Use app_name only for this referenced app.`
+    : "Target app: current app. Omit app_name in tool calls.";
+  return [
+    `User query: ${args.query}`,
+    targetText,
+    args.tsconfig_path ? `TypeScript config: ${args.tsconfig_path}` : "",
+    "",
+    `Intent: ${args.intent ?? "locate"}`,
+    "Explore with the available read-only tools. When you have enough evidence, call submit_report using observed candidate IDs.",
+  ]
+    .filter(Boolean)
+    .join("\n");
+}
