+++ src/prompts/local_agent_prompt.test.ts
@@ -12,20 +12,15 @@ describe("local_agent_prompt", () => {
       codeExplorerAvailable: true,
     });
     expect(prompt).toMatchSnapshot();
-    expect(prompt).toContain("Use `explore_code` first");
+    expect(prompt).toContain("use `explore_code` first");
     expect(prompt).toContain(
-      "Do not warm up with `list_files`, `grep`, or `read_file` before `explore_code`",
+      "do not warm up with `list_files`, `grep`, or `read_file` before it",
     );
     expect(prompt).toContain(
-      "include a compact code map that preserves every useful `path:start-end` reference",
-    );
-    expect(prompt).toContain("targeted_gap_search is a terminal bounded pass");
-    expect(prompt).toContain("run only the rendered Search targets");
-    expect(prompt).toContain(
-      "answer next from the report without `read_file`, `grep`, `list_files`, or another `explore_code` call",
+      "Follow the report's Action exactly as documented in the `explore_code` tool",
     );
     expect(prompt).toContain(
-      "read_targets means explain/locate requests should answer next from the report and cite the listed targets as jump points without `read_file`",
+      "do not call `explore_code` again for the same investigation",
     );
     expect(prompt).toContain(
       "When no authoritative explore_code report is available",
